Added different food types

+++ core/src/main/java/io/github/evolutionary_algorithm/Food.java
@@ -1,17 +1,17 @@
 package io.github.evolutionary_algorithm;
 
 public class Food {
-    private int nutrition;
+    private double nutrition;
     private int code;
     private int i;
     private int j;
-    public Food(int i,int j, int nutrition, int code){
+    public Food(int i,int j, double nutrition, int code){
         this.i = i;
         this.j = j;
         this.nutrition = nutrition;
         this.code = code;
     }
-    public int getNutrition() {
+    public double getNutrition() {
         return nutrition;
     }
